+++ lib/app/app.dart
@@ -3,14 +3,14 @@
 import 'package:flutter/material.dart';
 import '../prestars_exports.dart';
 
-
-
 class App extends StatelessWidget {
   final Environment env;
   const App({Key? key, required this.env}) : super(key: key);
   @override
   Widget build(BuildContext context) {
-    final routes = AppRoutes.routes();
+    final rootNavigatorKey = GlobalKey<NavigatorState>();
+    final shellNavigatorKey = GlobalKey<NavigatorState>();
+    final routes = AppRoutes.routes(rootNavigatorKey, shellNavigatorKey);
 
     return MaterialApp.router(
       scrollBehavior: MyCustomScrollBehavior(),