+++ lib/app/src/core/utils/_exports.dart
@@ -10,3 +10,4 @@ export 'my_custom_scroll_behavior.dart';
 export 'date_input_format.dart';
 export 'validators/_exports.dart';
 export 'debouncer.dart';
+export 'route_transition.dart';