+++ lib/app/src/modules/base/presenter/_exports.dart
@@ -2,3 +2,4 @@ export 'actions/_exports.dart';
 export 'components/_exports.dart';
 export 'controllers/_exports.dart';
 export 'base_page.dart';
+export 'default_page.dart';