+++ lib/app/src/modules/base/presenter/base_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import '../../../../../prestars_exports.dart';
 
 class BasePage extends StatefulWidget {
-  final Widget body;
+  final Widget child;
   final EdgeInsets? padding;
   final double? maxWidth;
   final double? maxHeight;
@@ -11,7 +11,7 @@ class BasePage extends StatefulWidget {
   final Widget? bottomWidget;
   const BasePage(
       {Key? key,
-      required this.body,
+      required this.child,
       this.maxWidth,
       this.maxHeight,
       this.padding,
@@ -33,31 +33,30 @@ class _BasePageState extends State<BasePage>
   void initState() {
     scrollController = ScrollController();
     baseNavigatorController = BaseNavigatorController();
+    baseNavigatorController.init();
     super.initState();
   }
 
   @override
   // ignore: must_call_super
   Widget build(BuildContext context) {
-    return WillPopScope(
-        onWillPop: () async => true,
-        child: Scaffold(
-          backgroundColor: ThemeService.colors.background,
-          appBar:
-              UiHeader.call(context, scrollController, baseNavigatorController),
-          body: SingleChildScrollView(
-            controller: scrollController,
-            child: Column(
-              children: [
-                Padding(
-                  padding: widget.padding ?? const EdgeInsets.all(16),
-                  child: widget.body,
-                ),
-                const UiFooter(),
-              ],
+    return Scaffold(
+      backgroundColor: ThemeService.colors.background,
+      appBar: UiHeader.call(context, scrollController, baseNavigatorController),
+      body: SizedBox(
+        height: context.height,
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Expanded(
+              child: widget.child,
             ),
-          ),
-        ));
+          ],
+        ),
+      ),
+    );
   }
 
   @override