+++ lib/app/src/modules/base/presenter/components/_exports.dart
@@ -1,2 +1,3 @@
 export 'ui_footer.dart';
 export 'ui_header.dart';
+export 'ui_navigator.dart';