+++ lib/app/src/modules/share/_exports.dart
@@ -2,3 +2,4 @@ export 'domain/_exports.dart';
 export 'infra/_exports.dart';
 export 'external/_exports.dart';
 export 'share_binds.dart';
+export 'presenter/_exports.dart';