+++ lib/app/src/modules/athlete/presenter/components/athlete_videos.dart
@@ -1,3 +1,4 @@
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:reactive_forms/reactive_forms.dart';
@@ -16,6 +17,9 @@ class AthleteVideos extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.start,
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
+          AutoSizeText(ConstantsLabels.videos,
+              style: ThemeService.styles.exo2Caption()),
+          const SizedBox(height: 20),
           Wrap(
             children: [
               SizedBox(
@@ -59,23 +63,51 @@ class AthleteVideos extends StatelessWidget {
                               return Padding(
                                 padding:
                                     const EdgeInsets.symmetric(horizontal: 8),
-                                child: InkWell(
-                                  onTap: videos.length <= 2
-                                      ? () => CaptureVideos.call(
-                                          form: controller.athleteForm)
-                                      : null,
-                                  child: Container(
-                                    height: 50,
-                                    width: 60,
-                                    decoration: BoxDecoration(
-                                        color: ThemeService.colors.white,
-                                        borderRadius: BorderRadius.circular(10),
-                                        border: Border.all(
-                                            color:
-                                                ThemeService.colors.primary)),
-                                    child: Center(
-                                        child: Text('Vídeo ${index + 1}')),
-                                  ),
+                                child: Stack(
+                                  children: [
+                                    InkWell(
+                                      onTap: videos.length <= 2
+                                          ? () => CaptureVideos.call(
+                                              form: controller.athleteForm)
+                                          : null,
+                                      child: Stack(
+                                        children: [
+                                          Container(
+                                            height: 60,
+                                            width: 60,
+                                            decoration: BoxDecoration(
+                                                color:
+                                                    ThemeService.colors.white,
+                                                borderRadius:
+                                                    BorderRadius.circular(10),
+                                                border: Border.all(
+                                                    color: ThemeService
+                                                        .colors.primary)),
+                                            child: Center(
+                                                child: Text(
+                                              'Vídeo ${index + 1}',
+                                              style: ThemeService.styles
+                                                  .exo2Body(),
+                                            )),
+                                          ),
+                                          Container(
+                                            height: 60,
+                                            width: 60,
+                                            alignment: Alignment.topRight,
+                                            child: InkWell(
+                                                onTap: () => controller
+                                                    .athleteForm
+                                                    .deleteVideo(index),
+                                                child: Icon(
+                                                  Icons.cancel_rounded,
+                                                  color: ThemeService
+                                                      .colors.primary,
+                                                )),
+                                          )
+                                        ],
+                                      ),
+                                    ),
+                                  ],
                                 ),
                               );
                             }