+++ changelog.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 26/07/2023
+
+### Changed
+- Removido . do final da frase Se cadastra aí e se torne um PRESTARS
+- Editado label do videosUrl
+- Alterado KeyBoardInput da data para number
+
+### Added
+
+- Adicionado botão remover aos videos
+- Adicionado aviso de desenvolvimento
+- Adicionado label dos vídeos
+
 ## [0.0.2] - 25/07/2023
 
 ### Fixed